move review day helpers to domain
